refactor: structure, update class overrides
