+++ src/core/Input/Checkbox/styles.css
