+++ src/core/Input/ToggleField/index.js
@@ -2,8 +2,8 @@
 
 import React from 'react'
 import PropTypes from 'prop-types'
-import { bemClassNames } from '../../utils'
-import './togglefield.css'
+import { bemClassNames } from '../../../utils'
+import './styles.css'
 
 const bem = bemClassNames('toggle-field')
