+++ src/core/Paper/styles.css
@@ -1,13 +1,12 @@
 /** @format */
 
 .d2ui-paper {
-    margin: 0 auto 0;
-    position: relative;
-    border-radius: 2px;
-    background-color: #ffffff;
-
-    &.padded {
+    &-defaults {
         padding: 2rem 2rem;
+        margin: 0 auto 0;
+        position: relative;
+        border-radius: 2px;
+        background-color: #ffffff;
     }
 
     &--elevation- {
